+++ rose-spring/src/main/java/io/zhijun/spring/core/env/ListenableConfigurableEnvironmentInitializer.java
@@ -4,17 +4,13 @@ import io.zhijun.spring.context.ConfigurableApplicationContextInitializer;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.core.env.ConfigurableEnvironment;
 
+import static io.zhijun.spring.constants.PropertyConstants.LISTENABLE_ENVIRONMENT_PROPERTY_NAME_PREFIX;
+
 /**
  * {@link ConfigurableApplicationContextInitializer} 实现，用于初始化 {@link ListenableConfigurableEnvironment}。
  * 配置属性：{@code rose.spring.listenable-environment.enabled}（默认 false）。
  */
 public class ListenableConfigurableEnvironmentInitializer extends ConfigurableApplicationContextInitializer {
-
-    public static final String PROPERTY_NAME_PREFIX = "rose.spring.listenable-environment.";
-
-    public static final String ENABLED_PROPERTY_NAME = PROPERTY_NAME_PREFIX + "enabled";
-
-    public static final boolean DEFAULT_ENABLED_PROPERTY_VALUE = false;
 
     @Override
     protected void initialize(ConfigurableApplicationContext context, ConfigurableEnvironment environment) {
@@ -27,11 +23,11 @@ public class ListenableConfigurableEnvironmentInitializer extends ConfigurableAp
 
     @Override
     public String getEnabledPropertyName() {
-        return ENABLED_PROPERTY_NAME;
+        return LISTENABLE_ENVIRONMENT_PROPERTY_NAME_PREFIX;
     }
 
     @Override
     public boolean getDefaultEnabled() {
-        return DEFAULT_ENABLED_PROPERTY_VALUE;
+        return false;
     }
 }